requirements.txt
streamlit==1.31.0
pydantic
sqlalchemy
plotly
statsmodels
streamlit_folium
folium
psycopg2-binary
streamlit-chat
st-annotated-text
streamlit-tags==1.2.8
st-pages==0.4.5

openai==1.30.2
streamlit-js-eval==0.1.7
streamlit-mic-recorder==0.0.8
pillow==10.3.0